backend/requirements.runtime.txt
fastapi==0.128.0
uvicorn==0.40.0
celery==5.6.0
redis==6.4.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
asyncpg==0.31.0
python-dotenv==1.2.1
loguru==0.7.3
requests==2.32.5
pydantic==2.12.5
python-jose==3.5.0
passlib==1.7.4
python-dateutil==2.9.0.post0
beautifulsoup4==4.14.3
lxml==6.0.2
numpy==2.4.0
pandas==2.3.3
